modelgalary/requirements.txt
scipy==0.19.0
django==2.0.4
requests==2.16.5
grpcio==1.3.5
numpy==1.12.1
tensorflow==1.7.0
pillow==4.1.1